Reverted potenital PII input logging
