+++ training/model_management/src/azureml/model/mgmt/utils/common_utils.py
@@ -17,6 +17,7 @@ from pathlib import Path
 from subprocess import PIPE, run, STDOUT
 from typing import Any, Dict, Tuple
 from applicationinsights import TelemetryClient
+from huggingface_hub.hf_api import HfApi, ModelInfo, ModelFilter
 
 tc = None
 
@@ -145,3 +146,15 @@ def tc_exception(e, message):
         tc.track_exception(value=e.__class__, properties={"exception": message})
     except Exception as e:
         print(f"Exception while logging exception to app insights: {e}")
+
+
+def check_model_id(model_id):
+    """Hugging face model info."""
+    try:
+        model_list: List[ModelInfo] = HfApi(endpoint=HF_ENDPOINT).list_models(filter=ModelFilter(model_name=model_id))
+        for info in model_list:
+            if model_id == info.modelId:
+                return True
+    except Exception as e:
+        raise ValueError(f"Failed to validate model id : {e}")
+    return False
